Refactor aspects data into a hook
