+++ src/containers/ContributeTabContainer/ContributeTabContainer.tsx
@@ -1,26 +1,14 @@
-import { FC, useMemo } from 'react';
+import { FC } from 'react';
 import { ApolloError } from '@apollo/client';
-import {
-  AspectCardFragment,
-  AuthorizationPrivilege,
-  CreateAspectOnContextInput,
-  Scalars,
-} from '../../models/graphql-schema';
-import { useApolloErrorHandler, useHub } from '../../hooks';
+import { AspectCardFragment, CreateAspectOnContextInput, Scalars } from '../../models/graphql-schema';
+import { useApolloErrorHandler, useHub, useAspects } from '../../hooks';
 import {
   AspectCardFragmentDoc,
-  useChallengeAspectsQuery,
   useCreateAspectFromContributeTabMutation,
   useDeleteAspectMutation,
-  useHubAspectsQuery,
-  useOpportunityAspectsQuery,
-  usePrivilegesOnChallengeContextQuery,
-  usePrivilegesOnHubContextQuery,
-  usePrivilegesOnOpportunityContextQuery,
 } from '../../hooks/generated/graphql';
 import { ContainerPropsWithProvided, renderComponentOrChildrenFn } from '../../utils/containers/ComponentOrChildrenFn';
 import removeFromCache from '../../utils/apollo-cache/removeFromCache';
-import useContextAspectCreatedSubscription from '../../domain/context/useContextAspectCreatedSubscription';
 
 export interface EntityIds {
   hubNameId: Scalars['UUID_NAMEID'];
@@ -53,127 +41,13 @@ const ContributeTabContainer: FC<ContributeContainerProps> = ({
   ...rendered
 }) => {
   const handleError = useApolloErrorHandler();
-  const { template, loading: hubLoading } = useHub();
+  const { template } = useHub();
 
-  const {
-    data: hubContextData,
-    loading: hubContextLoading,
-    error: hubContextError,
-    subscribeToMore: subscribeToHub,
-  } = usePrivilegesOnHubContextQuery({
-    variables: { hubNameId },
-    skip: !!(challengeNameId || opportunityNameId),
-    onError: handleError,
+  const { aspects, loading, error, canReadAspects, canCreateAspects, contextId } = useAspects({
+    hubNameId,
+    challengeNameId,
+    opportunityNameId,
   });
-  const hubContext = hubContextData?.hub?.context;
-  const hubContextPrivileges = hubContext?.authorization?.myPrivileges;
-  const canReadHubContext = hubContextPrivileges?.includes(AuthorizationPrivilege.Read);
-  const canCreateAspectOnHub = hubContextPrivileges?.includes(AuthorizationPrivilege.CreateAspect);
-  const {
-    data: hubAspectData,
-    loading: hubAspectLoading,
-    error: hubAspectError,
-  } = useHubAspectsQuery({
-    variables: { hubNameId },
-    skip: !canReadHubContext || !!(challengeNameId || opportunityNameId),
-    onError: handleError,
-  });
-  const hubAspects = hubAspectData?.hub?.context?.aspects;
-
-  const {
-    data: challengeContextData,
-    loading: challengeContextLoading,
-    error: challengeContextError,
-  } = usePrivilegesOnChallengeContextQuery({
-    variables: { hubNameId, challengeNameId },
-    skip: !challengeNameId || !!opportunityNameId,
-    onError: handleError,
-  });
-  const challengeContext = challengeContextData?.hub?.challenge?.context;
-  const challengeContextPrivileges = challengeContext?.authorization?.myPrivileges;
-  const canReadChallengeContext = challengeContextPrivileges?.includes(AuthorizationPrivilege.Read);
-  const canCreateAspectOnChallenge = challengeContextPrivileges?.includes(AuthorizationPrivilege.CreateAspect);
-
-  const {
-    data: challengeAspectData,
-    loading: challengeAspectLoading,
-    error: challengeAspectError,
-    subscribeToMore: subscribeToChallenges,
-  } = useChallengeAspectsQuery({
-    variables: { hubNameId, challengeNameId },
-    skip: !canReadChallengeContext || !challengeNameId || !!opportunityNameId,
-    onError: handleError,
-  });
-  const challengeAspects = challengeAspectData?.hub?.challenge?.context?.aspects;
-
-  const {
-    data: opportunityContextData,
-    loading: opportunityContextLoading,
-    error: opportunityContextError,
-  } = usePrivilegesOnOpportunityContextQuery({
-    variables: { hubNameId, opportunityNameId },
-    skip: !opportunityNameId,
-    onError: handleError,
-  });
-  const opportunityContext = opportunityContextData?.hub?.opportunity?.context;
-  const opportunityContextPrivileges = opportunityContext?.authorization?.myPrivileges;
-  const canReadOpportunityContext = opportunityContextPrivileges?.includes(AuthorizationPrivilege.Read);
-  const canCreateAspectOnOpportunity = opportunityContextPrivileges?.includes(AuthorizationPrivilege.CreateAspect);
-
-  const {
-    data: opportunityAspectData,
-    loading: opportunityAspectLoading,
-    error: opportunityAspectError,
-    subscribeToMore: subscribeToOpportunity,
-  } = useOpportunityAspectsQuery({
-    variables: { hubNameId, opportunityNameId },
-    skip: !canReadOpportunityContext || !opportunityNameId,
-    onError: handleError,
-    fetchPolicy: 'network-only',
-    nextFetchPolicy: 'cache-first',
-  });
-  const opportunityAspects = opportunityAspectData?.hub?.opportunity?.context?.aspects;
-
-  useContextAspectCreatedSubscription(hubAspectData, hubData => hubData?.hub?.context, subscribeToHub);
-  useContextAspectCreatedSubscription(
-    challengeAspectData,
-    challengeData => challengeData?.hub?.challenge?.context,
-    subscribeToChallenges
-  );
-  useContextAspectCreatedSubscription(
-    opportunityAspectData,
-    opportunityData => opportunityData?.hub?.opportunity?.context,
-    subscribeToOpportunity
-  );
-
-  const aspects: AspectWithPermissions[] | undefined = useMemo(
-    () =>
-      (hubAspects ?? challengeAspects ?? opportunityAspects)?.map(x => ({
-        ...x,
-        canDelete: x?.authorization?.myPrivileges?.includes(AuthorizationPrivilege.Delete),
-      })),
-    [hubAspects, challengeAspects, opportunityAspects]
-  );
-  const loading =
-    hubContextLoading ||
-    challengeContextLoading ||
-    opportunityContextLoading ||
-    hubAspectLoading ||
-    challengeAspectLoading ||
-    opportunityAspectLoading ||
-    hubLoading;
-  const error =
-    hubContextError ??
-    challengeContextError ??
-    opportunityContextError ??
-    hubAspectError ??
-    challengeAspectError ??
-    opportunityAspectError;
-
-  const canReadAspects = canReadHubContext ?? canReadChallengeContext ?? canReadOpportunityContext ?? true;
-  const canCreateAspects = canCreateAspectOnHub ?? canCreateAspectOnChallenge ?? canCreateAspectOnOpportunity ?? false;
-
-  const contextId = hubContext?.id ?? challengeContext?.id ?? opportunityContext?.id;
 
   const aspectTypes = template.aspectTemplates.map(x => x.type);
