+++ src/hooks/index.ts
@@ -22,6 +22,7 @@ export * from './useHub';
 export * from './useOrganization';
 export * from './useChallenge';
 export * from './useOpportunity';
+export * from './useAspects';
 
 export * from './useUrlParams';
 export * from './useCurrentBreakpoint';
