+++ src/pages/aspect/AspectSettingsPage.tsx
@@ -77,6 +77,7 @@ const AspectSettingsPage: FC<AspectSettingsPageProps> = ({ paths: _paths }) => {
               edit
               loading={state.loading || state.updating}
               aspect={toAspectFormInput(entities.aspect)}
+              aspectNames={entities.aspects?.map(x => x.displayName)}
               onChange={setAspect}
               onAddReference={actions.handleAddReference}
               onRemoveReference={actions.handleRemoveReference}
